+++ druid-shell/src/platform/x11/screen.rs
@@ -17,6 +17,6 @@
 use crate::screen::Monitor;
 
 pub(crate) fn get_monitors() -> Vec<Monitor> {
-    log::warn!("Screen::get_monitors() is currently unimplemented for X11 platforms.");
+    tracing::warn!("Screen::get_monitors() is currently unimplemented for X11 platforms.");
     Vec::new()
 }