+++ druid/examples/layout.rs
@@ -69,7 +69,7 @@ pub fn main() {
     let window = WindowDesc::new(build_app()).title("Very flexible");
 
     AppLauncher::with_window(window)
-        .use_simple_logger()
+        .use_env_tracing()
         .launch(0)
         .expect("launch failed");
 }