+++ CHANGELOG.md
@@ -8,6 +8,8 @@ You can find its changes [documented below](#070---2021-01-01).
 ### Highlights
 
 ### Added
+- Add `scroll()` method in WidgetExt ([#1600] by [@totsteps])
+- `write!` for `RichTextBuilder` ([#1596] by [@Maan2003])
 - Sub windows: Allow opening windows that share state with arbitrary parts of the widget hierarchy ([#1254] by [@rjwittams])
 - WindowCloseRequested/WindowDisconnected event when a window is closing ([#1254] by [@rjwittams])
 - RichTextBuilder ([#1520] by [@Maan2003])
@@ -17,11 +19,13 @@ You can find its changes [documented below](#070---2021-01-01).
 - Contexts: to_window and to_screen (useful for relatively positioning sub windows) ([#1532] by [@rjwittams])
 - WindowSizePolicy: allow windows to be sized by their content ([#1532] by [@rjwittams])
 - Implemented `Data` for more datatypes from `std` ([#1534] by [@derekdreery])
+- Shell: windows implementation from content_insets ([#1592] by [@HoNile])
 
 ### Changed
 
 - Warn on unhandled Commands ([#1533] by [@Maan2003])
 - `WindowDesc::new` takes the root widget directly instead of a closure ([#1559] by [@lassipulkkinen])
+- Switch to trace-based logging ([#1562] by [@PoignardAzur])
 
 ### Deprecated
 
@@ -413,6 +417,8 @@ Last release without a changelog :(
 [@derekdreery]: https://github.com/derekdreery
 [@MaximilianKoestler]: https://github.com/MaximilianKoestler
 [@lassipulkkinen]: https://github.com/lassipulkkinen
+[@Poignardazur]: https://github.com/PoignardAzur
+[@HoNile]: https://github.com/HoNile
 
 [#599]: https://github.com/linebender/druid/pull/599
 [#611]: https://github.com/linebender/druid/pull/611
@@ -611,8 +617,13 @@ Last release without a changelog :(
 [#1534]: https://github.com/linebender/druid/pull/1534
 [#1254]: https://github.com/linebender/druid/pull/1254
 [#1559]: https://github.com/linebender/druid/pull/1559
+[#1562]: https://github.com/linebender/druid/pull/1562
+[#1592]: https://github.com/linebender/druid/pull/1592
+[#1596]: https://github.com/linebender/druid/pull/1596
+[#1600]: https://github.com/linebender/druid/pull/1600
 
-[Unreleased]: https://github.com/linebender/druid/compare/v0.6.0...master
+[Unreleased]: https://github.com/linebender/druid/compare/v0.7.0...master
+[0.7.0]: https://github.com/linebender/druid/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/linebender/druid/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/linebender/druid/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/linebender/druid/compare/v0.3.2...v0.4.0